rag-service/requirements.txt
# rag-service/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
transformers
pathway